backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
opencv-python>=4.10.0
face-recognition>=1.3.0
dlib>=19.24.0
numpy>=1.24.0
pymongo>=4.6.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pandas>=2.0.0
